+++ app/src/main/java/com/hope/leadmap/mapdemo/sevenparameter/ProjectInfo.java
@@ -4,14 +4,14 @@ package com.hope.leadmap.mapdemo.sevenparameter;
  * Created by Hope on 2015/12/25.
  */
 public class ProjectInfo {
-    private static int middleline;
+    private static double middleline;
     private static int zonewide;
 
-    public static int getMiddleline() {
+    public static double getMiddleline() {
         return middleline;
     }
 
-    public static void setMiddleline(int middleline) {
+    public static void setMiddleline(double middleline) {
         ProjectInfo.middleline = middleline;
     }
